requirements.txt
bip-utils
tronpy
requests
python-dotenv
pandas
plotly
google-api-python-client